requirements.txt
pystray
requests
win10toast
pillow